requirements.txt
pysnmp-lextudio
snmpsim-lextudio
sv-ttk
Pillow